Ensure globals are functions before running `instanceof` checks against them.
`File` and `Blob` don't exist in Node, and `Buffer` and `ArrayBuffer` don't exist in HTML5.  So if an app uses those globals for some other purpose and also uses `has-binary`, it can crash with a "TypeError: Expecting a function in instanceof check, but got <whatever>" error.  This commit fixes that issue.